Comp/Cline/whisperApp/requirements.txt
streamlit
openai-whisper
torch
numpy
